Fix missing LogicalMin/LogicalMax
The global merge optimization pass initialized LogicalMin/LogicaMax to
zero, and thus if the actual value was zero it would be
removed from the output hid descriptor. Here we instead initialize it to
None causing it to be emitted on first use and then no longer emitted.

This is in accordance with the HID specification that specifies that a
LogicalMin/LogicalMax value is mandatory in the descriptor.